Use factory method for session_manager
Also add some abstraction over the SQL to allow for different SQL
dialects
